Explicitly use int64 to avoid int overflow on 32-bit
In order to build Ente on ARM, I had to switch from untyped int to int64, as it
seems untyped int defaults to int32 in Go.

Fixes:
```
ente/user.go:98:17: ExpectedKDFStrength (untyped int constant 4294967296) overflows int
ente/user.go:120:17: ExpectedKDFStrength (untyped int constant 4294967296) overflows int
../../pkg/controller/email/storage_warning.go:628:36: cannot use storageWarningOverageThreshold (untyped int constant 26843545600) as int value in map literal (overflows)
../../pkg/controller/emergency/recovery.go:131:23: cannot use 1000 * 1000 * 24 * 60 * 60 (untyped int constant 86400000000) as int value in assignment (overflows)
```

+++ server/pkg/controller/emergency/recovery.go
@@ -128,7 +128,7 @@ func (c *Controller) SendRecoveryReminder() {
 		return
 	}
 	log.Info(fmt.Sprintf("Found %d recovery rows", len(*rows)))
-	microsecondsInDay := 1000 * 1000 * 24 * 60 * 60
+	microsecondsInDay := int64(1000 * 1000 * 24 * 60) * 60
 	for _, row := range *rows {
 		logger := log.WithFields(log.Fields{
 			"userID":         row.UserID,
